# Golang/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Astro Wars Backend API",
        "title": "Astro Wars API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/account/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a user account by their ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Delete Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID. Required parameter to identify the user.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/forget-password": {
            "post": {
                "description": "Initiates password recovery (Currently placeholder).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Forget Password",
                "parameters": [
                    {
                        "description": "User Email. 'email' field is required.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.forgetPasswordRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/guest": {
            "post": {
                "description": "Creates a temporary guest account. Does not require a request body. Returns a JWT token and the guest user's data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Guest Login",
                "responses": {}
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticates a user using their username and password. Returns a JWT token and the user's data upon successful authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login Credentials. Both 'username' and 'password' are required.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.loginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates the user session. (Implementation depends on token handling logic, currently returns success message).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User Logout",
                "responses": {}
            }
        },
        "/auth/register": {
            "post": {
                "description": "Creates a new user account with the provided details. Returns a JWT token and the created user's data upon successful registration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "Registration Details. 'fullname', 'username', 'password', 'college', 'department', and 'academic_year' (1-5) are required. 'gdg_track' is optional.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.registerRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "handlers.forgetPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "User's email address (Required)",
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "handlers.loginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "User's password (Required)",
                    "type": "string",
                    "example": "P@ssw0rd123"
                },
                "username": {
                    "description": "User's username (Required)",
                    "type": "string",
                    "example": "mostafa_sensei"
                }
            }
        },
        "handlers.registerRequest": {
            "type": "object",
            "required": [
                "academic_year",
                "college",
                "department",
                "fullname",
                "password",
                "username"
            ],
            "properties": {
                "academic_year": {
                    "description": "Academic year from 1 to 5 (Required)",
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1,
                    "example": 3
                },
                "college": {
                    "description": "User's college (Required)",
                    "type": "string",
                    "example": "Computers \u0026 Info"
                },
                "department": {
                    "description": "User's department (Required)",
                    "type": "string",
                    "example": "Computer Science"
                },
                "fullname": {
                    "description": "Full Name (Required)",
                    "type": "string",
                    "example": "Mostafa Ahmed"
                },
                "gdg_track": {
                    "description": "GDG Track name (Optional, can be null)",
                    "type": "string",
                    "x-nullable": true,
                    "example": "Flutter"
                },
                "password": {
                    "description": "Password for login (Required)",
                    "type": "string",
                    "example": "P@ssw0rd123"
                },
                "username": {
                    "description": "Username for login (Required)",
                    "type": "string",
                    "example": "mostafa_sensei"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}